third_party/blink/web_tests/external/wpt/html/canvas/historical.window-expected.txt

This is a testharness.js-based test.
[FAIL] CanvasRenderingContext2D.scrollPathIntoView method is removed
  assert_equals: expected (undefined) undefined but got (function) function "function scrollPathIntoView() { [native code] }"
Harness: the test ran to completion.
